refector: svg bound 계산 로직 분리
